Refactor emailService into commonFormService - The email function is technically a common service call that can be used by any form - Better to consolidate "common" functionality with similar shapes to the same service
