+++ app/frontend/src/services/index.js
@@ -1,2 +1,2 @@
-export { default as emailService } from './emailService';
+export { default as commonFormService } from './commonFormService';
 export { default as minesOperatorScreeningService } from './minesOperatorScreeningService';
